requirements.txt
openai
flask
PyPDF
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client==2.2.4
python-dotenv
langchain==0.0.225

-e .